requirements.txt
pywebview>=6.0
requests>=2.32.0
